feat(UI-973): re-write sidebar to use popoverList
