+++ src/components/molecules/index.ts
@@ -9,7 +9,7 @@ export { ImageMotion } from "@components/molecules/imageMotion";
 export { MapMenu } from "@components/molecules/mapMenu";
 export { Menu } from "@components/molecules/menu";
 export { Modal } from "@components/molecules/modal";
-export { Popover, PopoverContent, PopoverTrigger } from "@components/molecules/popover";
+export { Popover, PopoverContentBase, PopoverCloseButton, PopoverTrigger } from "@components/molecules/popover";
 export { RefreshButton } from "@components/molecules/refreshButton";
 export { Select } from "@components/molecules/select";
 export { SortButton } from "@components/molecules/sortButton";
