+++ src/components/molecules/popover/utilities/mergeRefs.ts
@@ -1,0 +1,5 @@
+import { useMergeRefs } from "@floating-ui/react";
+
+export const useMergeRefsCustom = (...refs: any[]) => {
+	return useMergeRefs(refs.filter(Boolean));
+};
