+++ src/components/organisms/sidebar/sidebar.tsx
@@ -12,8 +12,9 @@ import { useLoggerStore, useUserStore } from "@store";
 
 import { Badge, Button, IconSvg, Loader } from "@components/atoms";
 import { MenuToggle } from "@components/atoms/menuToggle";
-import { Popover, PopoverContent, PopoverTrigger } from "@components/molecules";
+import { Popover, PopoverTrigger } from "@components/molecules";
 import { Menu } from "@components/molecules/menu";
+import { PopoverContent } from "@components/molecules/popover/index";
 
 import { IconLogo, IconLogoName } from "@assets/image";
 import { FileIcon, HelpIcon, ListDetailsIcon } from "@assets/image/icons";
