+++ src/hooks/index.ts
@@ -8,5 +8,4 @@ export { useSort } from "@hooks/useSort";
 export { useToastAndLog } from "@hooks/useToastAndLog";
 export { useVirtualizedList } from "@hooks/useVirtualizedList";
 export { useWindowDimensions } from "@hooks/useWindowDimensions";
-export { usePopover } from "@src/hooks/usePopover";
-export { usePopoverList } from "@src/hooks/usePopoverList";
+export { usePopover, usePopoverList } from "@src/hooks/usePopover";
