+++ src/types/components/popover.type.ts
@@ -1,4 +1,4 @@
 import { usePopover, usePopoverList } from "@src/hooks";
 
-export type PopopverContextType = ReturnType<typeof usePopover> | null;
-export type PopopverListContextType = ReturnType<typeof usePopoverList> | null;
+export type PopoverContextType = ReturnType<typeof usePopover> | null;
+export type PopoverListContextType = ReturnType<typeof usePopoverList> | null;
